organize stories from builder
