--- packages/ui/stories/FormGenerator.data.tsx
@@ -1,213 +1,0 @@
-export const formData = {
-  properties: {
-    clientSecret: {
-      componentProperty: true,
-      deprecated: false,
-      description: 'Client Secret received from Concur upon registration.',
-      displayName: 'Client Secret',
-      group: 'producer',
-      javaType: 'java.lang.String',
-      kind: 'property',
-      label: 'producer',
-      required: true,
-      secret: true,
-      type: 'string',
-      tags: ['oauth-client-secret'],
-      order: 3,
-    },
-    authorizationEndpoint: {
-      componentProperty: true,
-      defaultValue: 'https://us.api.concursolutions.com/oauth2/v0/authorize',
-      deprecated: false,
-      description:
-        'Authorization URL consists of base URL and ends with <code>/authorize</code>. Base URL differs based on geolocation and should be received from Concur upon registration.',
-      displayName: 'Authorization URL',
-      group: 'producer',
-      javaType: 'java.lang.String',
-      kind: 'property',
-      label: 'producer',
-      required: true,
-      secret: false,
-      type: 'string',
-      tags: ['oauth-authorization-url'],
-      order: 6,
-    },
-    clientId: {
-      componentProperty: true,
-      deprecated: false,
-      description: 'Client ID received from Concur upon registration.',
-      displayName: 'Client ID',
-      group: 'producer',
-      javaType: 'java.lang.String',
-      kind: 'property',
-      label: 'producer',
-      required: true,
-      secret: false,
-      type: 'string',
-      tags: ['oauth-client-id'],
-      order: 2,
-    },
-    host: {
-      componentProperty: true,
-      defaultValue: 'https://us.api.concursolutions.com',
-      deprecated: false,
-      description:
-        'API Endpoint used to access Concur instance in the form of https://hostname:port. This value depends on the geolocation received from Concur upon registration.',
-      displayName: 'API Endpoint',
-      group: 'producer',
-      javaType: 'java.lang.String',
-      kind: 'property',
-      label: 'producer',
-      required: true,
-      secret: false,
-      type: 'string',
-      tags: ['host'],
-      order: 10,
-    },
-    tokenEndpoint: {
-      componentProperty: true,
-      defaultValue: 'https://us.api.concursolutions.com/oauth2/v0/token',
-      deprecated: false,
-      description:
-        'Token URL consists of base URL and ends with <code>/token</code>. Base URL differs based on geolocation and should be received from Concur upon registration.',
-      displayName: 'Token URL',
-      group: 'producer',
-      javaType: 'java.lang.String',
-      kind: 'property',
-      label: 'producer',
-      required: true,
-      secret: false,
-      type: 'string',
-      tags: ['oauth-access-token-url'],
-      order: 7,
-    },
-    oauthScopes: {
-      componentProperty: true,
-      defaultValue:
-        'ATTEND CONFIG CONREQ ERECPT EVS EXPRPT CCARD BANK EXTRCT FISVC FOP GHOST IMAGE INSGHT INVPMT INVPO INVTV INVVEN ITINER JOBLOG LIST MTNG NOTIF PAYBAT RCTIMG SUPSVC TAXINV TRVPRF PASSV EMERG TSAI TMCSP MEDIC UNUTX TRVPTS TRVREQ TWS USER COMPANY',
-      deprecated: false,
-      description:
-        'Scopes required as a list of space-delimited, case-sensitive strings. Scopes limit access and do not grant any additional permission beyond that which the user already has. Consult the Concur API documentation on the available Scopes.',
-      displayName: 'Scopes',
-      group: 'producer',
-      javaType: 'java.lang.String',
-      kind: 'property',
-      label: 'producer',
-      required: true,
-      secret: false,
-      type: 'string',
-      tags: ['oauth-scope'],
-      order: 8,
-    },
-    test01: {
-      type: 'color',
-      displayName: 'Color',
-      description: 'This is a native color picker widget',
-    },
-    test02: {
-      type: 'date',
-      displayName: 'Date',
-      description: 'This is a date input',
-    },
-    test03: {
-      type: 'datetime-local',
-      displayName: 'DateTime-Local',
-      description: 'This is a datetime-local input',
-    },
-    test04: {
-      type: 'duration',
-      displayName: 'Duration',
-      description: 'This is a duration input',
-    },
-    test05: {
-      type: 'email',
-      displayName: 'Email',
-      description: 'This is an email input',
-    },
-    test06: {
-      type: 'file',
-      displayName: 'File',
-      description: 'This is a file input',
-    },
-    test07: {
-      type: 'hidden',
-      displayName: 'Hidden',
-      description: 'This is a hidden input',
-    },
-    test08: {
-      type: 'month',
-      displayName: 'Month',
-      description: 'This is a native month input',
-    },
-    test09: {
-      type: 'password',
-      displayName: 'Password',
-      description: 'This is a native password input',
-    },
-    test10: {
-      type: 'range',
-      displayName: 'Range',
-      description: 'This is a native range input',
-    },
-    test11: {
-      type: 'tel',
-      displayName: 'Tel',
-      description: 'This is a native tel input',
-    },
-    test12: {
-      type: 'url',
-      displayName: 'URL',
-      description: 'This is a url input',
-    },
-    test13: {
-      type: 'week',
-      displayName: 'Week',
-      description: 'This is a week input',
-    },
-    test14: {
-      type: 'search',
-      displayName: 'Search',
-      description: 'This is a search input',
-    },
-    test15: {
-      type: 'textarea',
-      displayName: 'Textarea',
-      description: 'This is a textarea input',
-    },
-    test16: {
-      type: 'time',
-      displayName: 'Time',
-      description: 'This is a time input',
-    },
-    test17: {
-      type: 'boolean',
-      displayName: 'Boolean',
-      description: 'This is a boolean input',
-    },
-    test18: {
-      type: 'checkbox',
-      displayName: 'Checkbox',
-      description: 'This is a checkbox input',
-    },
-    test19: {
-      type: 'int',
-      displayName: 'Int',
-      description: 'This is a int input',
-    },
-    test20: {
-      type: 'integer',
-      displayName: 'Integer',
-      description: 'This is a integer input',
-    },
-    test21: {
-      type: 'string',
-      displayName: 'String',
-      description: 'This is a string input',
-    },
-    test22: {
-      type: 'select',
-      displayName: 'Select',
-      description: 'This is a select input',
-    },
-  },
-};
